spotify_analysis/requirements.txt
async-timeout==5.0.1
certifi==2025.4.26
charset-normalizer==3.4.2
contourpy==1.1.1
cycler==0.12.1
fonttools==4.57.0
idna==3.10
importlib-resources==6.4.5
kiwisolver==1.4.7
matplotlib==3.7.5
mysql-connector-python==9.0.0
numpy==1.24.4
packaging==25.0
pandas==2.0.3
pillow==10.4.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
pytz==2025.2
redis==6.1.0
requests==2.32.3
six==1.17.0
spotipy==2.25.1
tzdata==2025.2
urllib3==2.2.3
zipp==3.20.2
